+++ apps/test/unit/p5lab/GameLabSpriteTest.js
@@ -6,7 +6,6 @@ import {
   sandboxDocumentBody
 } from '../../util/testUtils';
 import createP5Wrapper, {
-  createStatefulP5Wrapper,
   expectAnimationsAreClones
 } from '../../util/gamelab/TestableP5Wrapper';
 
@@ -698,618 +697,6 @@ describe('P5SpriteWrapper', function() {
     });
   });
 
-  describe('collisions with groups', function() {
-    let sprite, spriteTarget1, spriteTarget2, group;
-    let p5WrapperStateful, createStatefulGroup, createStatefulSprite;
-
-    beforeEach(function() {
-      p5WrapperStateful = createStatefulP5Wrapper();
-      createStatefulGroup = p5WrapperStateful.p5.createGroup.bind(
-        p5WrapperStateful.p5
-      );
-      createStatefulSprite = p5WrapperStateful.p5.createSprite.bind(
-        p5WrapperStateful.p5
-      );
-      spriteTarget1 = createStatefulSprite(0, 0, 100, 100);
-      spriteTarget2 = createStatefulSprite(400, 400, 100, 100);
-      sprite = createStatefulSprite(200, 200, 100, 100);
-      group = createStatefulGroup();
-      group.add(spriteTarget1);
-      group.add(spriteTarget2);
-    });
-
-    it('isTouching returns false when sprite touches nothing in the group', function() {
-      let result;
-      for (let i = 0; i < group.length + 1; i++) {
-        result = sprite.isTouching(group);
-      }
-      expect(result).to.equal(false);
-    });
-
-    it('isTouching returns true when sprite touches anything in the group', function() {
-      spriteTarget1.x = sprite.x;
-      spriteTarget1.y = sprite.y;
-
-      let result2;
-      for (let i = 0; i < group.length + 1; i++) {
-        result2 = sprite.isTouching(group);
-      }
-      expect(result2).to.equal(true);
-    });
-
-    it('collide returns false when sprite touches nothing in the group', function() {
-      let result;
-      for (let i = 0; i < group.length + 1; i++) {
-        result = sprite.collide(group);
-      }
-      expect(result).to.equal(false);
-    });
-
-    it('collide returns true when sprite touches anything in the group', function() {
-      spriteTarget1.x = sprite.x;
-      spriteTarget1.y = sprite.y;
-
-      let result2;
-      for (let i = 0; i < group.length + 1; i++) {
-        result2 = sprite.collide(group);
-      }
-      expect(result2).to.equal(true);
-    });
-
-    it('bounce returns false when sprite touches nothing in the group', function() {
-      let result;
-      for (let i = 0; i < group.length + 1; i++) {
-        result = sprite.bounce(group);
-      }
-      expect(result).to.equal(false);
-    });
-
-    it('bounce returns true when sprite touches anything in the group', function() {
-      spriteTarget1.x = sprite.x;
-      spriteTarget1.y = sprite.y;
-
-      let result2;
-      for (let i = 0; i < group.length + 1; i++) {
-        result2 = sprite.bounce(group);
-      }
-      expect(result2).to.equal(true);
-    });
-
-    it('bounceOff returns false when sprite touches nothing in the group', function() {
-      let result;
-      for (let i = 0; i < group.length + 1; i++) {
-        result = sprite.bounceOff(group);
-      }
-      expect(result).to.equal(false);
-    });
-
-    it('bounceOff returns true when sprite touches anything in the group', function() {
-      spriteTarget1.x = sprite.x;
-      spriteTarget1.y = sprite.y;
-
-      let result2;
-      for (let i = 0; i < group.length + 1; i++) {
-        result2 = sprite.bounceOff(group);
-      }
-      expect(result2).to.equal(true);
-    });
-
-    it('displace returns false when sprite touches nothing in the group', function() {
-      let result;
-      for (let i = 0; i < group.length + 1; i++) {
-        result = sprite.displace(group);
-      }
-      expect(result).to.equal(false);
-    });
-
-    it('displace returns true when sprite touches anything in the group', function() {
-      spriteTarget1.x = sprite.x;
-      spriteTarget1.y = sprite.y;
-
-      let result2;
-      for (let i = 0; i < group.length + 1; i++) {
-        result2 = sprite.displace(group);
-      }
-      expect(result2).to.equal(true);
-    });
-
-    it('overlap returns false when sprite touches nothing in the group', function() {
-      let result;
-      for (let i = 0; i < group.length + 1; i++) {
-        result = sprite.overlap(group);
-      }
-      expect(result).to.equal(false);
-    });
-
-    it('overlap returns true when sprite touches anything in the group', function() {
-      spriteTarget1.x = sprite.x;
-      spriteTarget1.y = sprite.y;
-
-      let result2;
-      for (let i = 0; i < group.length + 1; i++) {
-        result2 = sprite.overlap(group);
-      }
-      expect(result2).to.equal(true);
-    });
-  });
-
-  describe('sprite.collide(sprite)', function() {
-    var SIZE = 10;
-    var pInst;
-    var spriteA, spriteB;
-    var callCount, pairs;
-
-    function testCallback(a, b) {
-      callCount++;
-      pairs.push([a.name, b.name]);
-    }
-
-    function moveAToB(a, b) {
-      a.position.x = b.position.x;
-    }
-
-    beforeEach(function() {
-      pInst = new p5(function() {});
-      callCount = 0;
-      pairs = [];
-
-      function createTestSprite(letter, position) {
-        var sprite = pInst.createSprite(position, 0, SIZE, SIZE);
-        sprite.name = 'sprite' + letter;
-        return sprite;
-      }
-
-      spriteA = createTestSprite('A', 2 * SIZE);
-      spriteB = createTestSprite('B', 4 * SIZE);
-
-      // Assert initial test state:
-      // - Two total sprites
-      // - no two sprites overlap
-      expect(pInst.allSprites.length).to.equal(2);
-      pInst.allSprites.forEach(function(caller) {
-        pInst.allSprites.forEach(function(callee) {
-          expect(caller.overlap(callee)).to.be.false;
-        });
-      });
-    });
-
-    afterEach(function() {
-      pInst.remove();
-    });
-
-    it('false if sprites do not overlap', function() {
-      expect(spriteA.collide(spriteB)).to.be.false;
-      expect(spriteB.collide(spriteA)).to.be.false;
-    });
-
-    it('true if sprites overlap', function() {
-      moveAToB(spriteA, spriteB);
-      expect(spriteA.collide(spriteB)).to.be.true;
-
-      moveAToB(spriteA, spriteB);
-      expect(spriteB.collide(spriteA)).to.be.true;
-    });
-
-    it('calls callback once if sprites overlap', function() {
-      expect(callCount).to.equal(0);
-
-      moveAToB(spriteA, spriteB);
-      spriteA.collide(spriteB, testCallback);
-      expect(callCount).to.equal(1);
-
-      moveAToB(spriteA, spriteB);
-      spriteB.collide(spriteA, testCallback);
-      expect(callCount).to.equal(2);
-    });
-
-    it('does not call callback if sprites do not overlap', function() {
-      expect(callCount).to.equal(0);
-      spriteA.collide(spriteB, testCallback);
-      expect(callCount).to.equal(0);
-      spriteB.collide(spriteA, testCallback);
-      expect(callCount).to.equal(0);
-    });
-
-    describe('passes collider and collidee to callback', function() {
-      it('A-B', function() {
-        moveAToB(spriteA, spriteB);
-        spriteA.collide(spriteB, testCallback);
-        expect(pairs).to.deep.equal([[spriteA.name, spriteB.name]]);
-      });
-
-      it('B-A', function() {
-        moveAToB(spriteA, spriteB);
-        spriteB.collide(spriteA, testCallback);
-        expect(pairs).to.deep.equal([[spriteB.name, spriteA.name]]);
-      });
-    });
-
-    it('does not reposition either sprite when sprites do not overlap', function() {
-      var initialPositionA = spriteA.position.copy();
-      var initialPositionB = spriteB.position.copy();
-
-      spriteA.collide(spriteB);
-
-      expectVectorsAreClose(spriteA.position, initialPositionA);
-      expectVectorsAreClose(spriteB.position, initialPositionB);
-
-      spriteB.collide(spriteA);
-
-      expectVectorsAreClose(spriteA.position, initialPositionA);
-      expectVectorsAreClose(spriteB.position, initialPositionB);
-    });
-
-    describe('displaces the caller out of collision when sprites do overlap', function() {
-      it('to the left', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-
-        var expectedPositionA = spriteB.position.copy().add(-SIZE, 0);
-        var expectedPositionB = spriteB.position.copy();
-
-        spriteA.collide(spriteB);
-
-        expectVectorsAreClose(spriteA.position, expectedPositionA);
-        expectVectorsAreClose(spriteB.position, expectedPositionB);
-      });
-
-      it('to the right', function() {
-        spriteA.position.x = spriteB.position.x + 1;
-
-        var expectedPositionA = spriteB.position.copy().add(SIZE, 0);
-        var expectedPositionB = spriteB.position.copy();
-
-        spriteA.collide(spriteB);
-
-        expectVectorsAreClose(spriteA.position, expectedPositionA);
-        expectVectorsAreClose(spriteB.position, expectedPositionB);
-      });
-
-      it('caller and callee reversed', function() {
-        spriteA.position.x = spriteB.position.x + 1;
-
-        var expectedPositionA = spriteA.position.copy();
-        var expectedPositionB = spriteA.position.copy().add(-SIZE, 0);
-
-        spriteB.collide(spriteA);
-
-        expectVectorsAreClose(spriteA.position, expectedPositionA);
-        expectVectorsAreClose(spriteB.position, expectedPositionB);
-      });
-    });
-
-    it('does not change velocity of either sprite when sprites do not overlap', function() {
-      var initialVelocityA = spriteA.velocity.copy();
-      var initialVelocityB = spriteB.velocity.copy();
-
-      spriteA.collide(spriteB);
-
-      expectVectorsAreClose(spriteA.velocity, initialVelocityA);
-      expectVectorsAreClose(spriteB.velocity, initialVelocityB);
-
-      spriteB.collide(spriteA);
-
-      expectVectorsAreClose(spriteA.velocity, initialVelocityA);
-      expectVectorsAreClose(spriteB.velocity, initialVelocityB);
-    });
-
-    describe('matches caller velocity to callee velocity when sprites do overlap', function() {
-      it('when callee velocity is zero', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-        spriteA.velocity.x = 2;
-        spriteB.velocity.x = 0;
-
-        var expectedVelocityA = spriteB.velocity.copy();
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteA.collide(spriteB);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-
-      it('when callee velocity is nonzero', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-        spriteA.velocity.x = 2;
-        spriteB.velocity.x = -1;
-
-        var expectedVelocityA = spriteB.velocity.copy();
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteA.collide(spriteB);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-
-      it('only along x axis when only displaced along x axis', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-        spriteA.velocity.x = 2;
-        spriteA.velocity.y = 3;
-        spriteB.velocity.x = -1;
-
-        // Expect to change velocity only along X
-        var expectedVelocityA = spriteA.velocity.copy();
-        expectedVelocityA.x = spriteB.velocity.x;
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteA.collide(spriteB);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-
-      it('only along y axis when only displaced along y axis', function() {
-        spriteA.position.x = spriteB.position.x;
-        spriteA.position.y = spriteB.position.y + 1;
-        spriteA.velocity.x = 2;
-        spriteA.velocity.y = 3;
-        spriteB.velocity.x = -1;
-        spriteB.velocity.y = 1.5;
-
-        // Expect to change velocity only along Y
-        var expectedVelocityA = spriteA.velocity.copy();
-        expectedVelocityA.y = spriteB.velocity.y;
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteA.collide(spriteB);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-
-      it('caller and callee reversed', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-        spriteA.velocity.x = 2;
-        spriteB.velocity.x = -1;
-
-        var expectedVelocityA = spriteA.velocity.copy();
-        var expectedVelocityB = spriteA.velocity.copy();
-
-        spriteB.collide(spriteA);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-    });
-  });
-
-  describe('sprite.displace(sprite)', function() {
-    var SIZE = 10;
-    var pInst;
-    var spriteA, spriteB;
-    var callCount, pairs;
-
-    function testCallback(a, b) {
-      callCount++;
-      pairs.push([a.name, b.name]);
-    }
-
-    function moveAToB(a, b) {
-      a.position.x = b.position.x;
-    }
-
-    beforeEach(function() {
-      pInst = new p5(function() {});
-      callCount = 0;
-      pairs = [];
-
-      function createTestSprite(letter, position) {
-        var sprite = pInst.createSprite(position, 0, SIZE, SIZE);
-        sprite.name = 'sprite' + letter;
-        return sprite;
-      }
-
-      spriteA = createTestSprite('A', 2 * SIZE);
-      spriteB = createTestSprite('B', 4 * SIZE);
-
-      // Assert initial test state:
-      // - Two total sprites
-      // - no two sprites overlap
-      expect(pInst.allSprites.length).to.equal(2);
-      pInst.allSprites.forEach(function(caller) {
-        pInst.allSprites.forEach(function(callee) {
-          expect(caller.overlap(callee)).to.be.false;
-        });
-      });
-    });
-
-    afterEach(function() {
-      pInst.remove();
-    });
-
-    it('false if sprites do not overlap', function() {
-      expect(spriteA.displace(spriteB)).to.be.false;
-      expect(spriteB.displace(spriteA)).to.be.false;
-    });
-
-    it('true if sprites overlap', function() {
-      moveAToB(spriteA, spriteB);
-      expect(spriteA.displace(spriteB)).to.be.true;
-
-      moveAToB(spriteA, spriteB);
-      expect(spriteB.displace(spriteA)).to.be.true;
-    });
-
-    it('calls callback once if sprites overlap', function() {
-      expect(callCount).to.equal(0);
-
-      moveAToB(spriteA, spriteB);
-      spriteA.displace(spriteB, testCallback);
-      expect(callCount).to.equal(1);
-
-      moveAToB(spriteA, spriteB);
-      spriteB.displace(spriteA, testCallback);
-      expect(callCount).to.equal(2);
-    });
-
-    it('does not call callback if sprites do not overlap', function() {
-      expect(callCount).to.equal(0);
-      spriteA.displace(spriteB, testCallback);
-      expect(callCount).to.equal(0);
-      spriteB.displace(spriteA, testCallback);
-      expect(callCount).to.equal(0);
-    });
-
-    describe('passes collider and collidee to callback', function() {
-      it('A-B', function() {
-        moveAToB(spriteA, spriteB);
-        spriteA.displace(spriteB, testCallback);
-        expect(pairs).to.deep.equal([[spriteA.name, spriteB.name]]);
-      });
-
-      it('B-A', function() {
-        moveAToB(spriteA, spriteB);
-        spriteB.displace(spriteA, testCallback);
-        expect(pairs).to.deep.equal([[spriteB.name, spriteA.name]]);
-      });
-    });
-
-    it('does not reposition either sprite when sprites do not overlap', function() {
-      var initialPositionA = spriteA.position.copy();
-      var initialPositionB = spriteB.position.copy();
-
-      spriteA.displace(spriteB);
-
-      expectVectorsAreClose(spriteA.position, initialPositionA);
-      expectVectorsAreClose(spriteB.position, initialPositionB);
-
-      spriteB.displace(spriteA);
-
-      expectVectorsAreClose(spriteA.position, initialPositionA);
-      expectVectorsAreClose(spriteB.position, initialPositionB);
-    });
-
-    describe('displaces the callee out of collision when sprites do overlap', function() {
-      it('to the left', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-
-        var expectedPositionA = spriteA.position.copy();
-        var expectedPositionB = spriteA.position.copy().add(SIZE, 0);
-
-        spriteA.displace(spriteB);
-
-        expectVectorsAreClose(spriteA.position, expectedPositionA);
-        expectVectorsAreClose(spriteB.position, expectedPositionB);
-      });
-
-      it('to the right', function() {
-        spriteA.position.x = spriteB.position.x + 1;
-
-        var expectedPositionA = spriteA.position.copy();
-        var expectedPositionB = spriteA.position.copy().add(-SIZE, 0);
-
-        spriteA.displace(spriteB);
-
-        expectVectorsAreClose(spriteA.position, expectedPositionA);
-        expectVectorsAreClose(spriteB.position, expectedPositionB);
-      });
-
-      it('caller and callee reversed', function() {
-        spriteA.position.x = spriteB.position.x + 1;
-
-        var expectedPositionA = spriteB.position.copy().add(SIZE, 0);
-        var expectedPositionB = spriteB.position.copy();
-
-        spriteB.displace(spriteA);
-
-        expectVectorsAreClose(spriteA.position, expectedPositionA);
-        expectVectorsAreClose(spriteB.position, expectedPositionB);
-      });
-    });
-
-    it('does not change velocity of either sprite when sprites do not overlap', function() {
-      var initialVelocityA = spriteA.velocity.copy();
-      var initialVelocityB = spriteB.velocity.copy();
-
-      spriteA.displace(spriteB);
-
-      expectVectorsAreClose(spriteA.velocity, initialVelocityA);
-      expectVectorsAreClose(spriteB.velocity, initialVelocityB);
-
-      spriteB.displace(spriteA);
-
-      expectVectorsAreClose(spriteA.velocity, initialVelocityA);
-      expectVectorsAreClose(spriteB.velocity, initialVelocityB);
-    });
-
-    describe('does not change callee velocity', function() {
-      it('when caller velocity is zero', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-        spriteA.velocity.x = 0;
-        spriteB.velocity.x = 2;
-
-        var expectedVelocityA = spriteA.velocity.copy();
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteA.displace(spriteB);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-
-      it('when caller velocity is nonzero', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-        spriteA.velocity.x = 2;
-        spriteB.velocity.x = -1;
-
-        var expectedVelocityA = spriteA.velocity.copy();
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteA.displace(spriteB);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-
-      it('when only displaced along x axis', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-        spriteA.velocity.x = 2;
-        spriteA.velocity.y = 0.5;
-        spriteB.velocity.x = -1;
-        spriteB.velocity.y = 3;
-
-        var expectedVelocityA = spriteA.velocity.copy();
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteA.displace(spriteB);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-
-      it('when only displaced along y axis', function() {
-        spriteA.position.x = spriteB.position.x;
-        spriteA.position.y = spriteB.position.y + 1;
-        spriteA.velocity.x = 2;
-        spriteA.velocity.y = 3;
-        spriteB.velocity.x = -1;
-        spriteB.velocity.y = 1.5;
-
-        var expectedVelocityA = spriteA.velocity.copy();
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteA.displace(spriteB);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-
-      it('caller and callee reversed', function() {
-        spriteA.position.x = spriteB.position.x - 1;
-        spriteA.velocity.x = 2;
-        spriteB.velocity.x = -1;
-
-        var expectedVelocityA = spriteA.velocity.copy();
-        var expectedVelocityB = spriteB.velocity.copy();
-
-        spriteB.displace(spriteA);
-
-        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
-        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
-      });
-    });
-  });
-
   function createTestAnimation(frameCount = 1, looping = true) {
     let image = new p5.Image(100, 100, p5Wrapper.p5);
     let frames = [];
@@ -1321,20 +708,5 @@ describe('P5SpriteWrapper', function() {
     animation.looping = looping;
     animation.frameDelay = 1;
     return animation;
-  }
-
-  function expectVectorsAreClose(vA, vB) {
-    var failMsg =
-      'Expected <' +
-      vA.x +
-      ', ' +
-      vA.y +
-      '> to equal <' +
-      vB.x +
-      ', ' +
-      vB.y +
-      '>';
-    expect(vA.x).to.be.closeTo(vB.x, 0.00001, failMsg);
-    expect(vA.y).to.be.closeTo(vB.y, 0.00001, failMsg);
   }
 });
