+++ apps/test/unit/p5lab/GameLabSpriteTestTwo.js
@@ -1,0 +1,715 @@
+/* @file Test of our p5.play Sprite wrapper object */
+/* global p5 */
+import {expect} from '../../util/reconfiguredChai';
+import {sandboxDocumentBody} from '../../util/testUtils';
+import createP5Wrapper, {
+  createStatefulP5Wrapper
+} from '../../util/gamelab/TestableP5Wrapper';
+
+describe('P5SpriteWrapper', function() {
+  let p5Wrapper, createSprite;
+
+  // Using the aggressive sandbox here because the P5 library generates
+  // a default canvas when it's not attached to an existing one.
+  sandboxDocumentBody();
+
+  beforeEach(function() {
+    p5Wrapper = createP5Wrapper();
+    createSprite = p5Wrapper.p5.createSprite.bind(p5Wrapper.p5);
+  });
+
+  describe('property aliases', function() {
+    let testSprite;
+
+    beforeEach(function() {
+      testSprite = createSprite();
+    });
+
+    it('aliases position.x to positionX', function() {
+      testSprite.position.x = 1;
+      expect(testSprite.position.x).to.equal(testSprite.x);
+      const newValue = 2;
+      testSprite.x = newValue;
+      expect(testSprite.position.x)
+        .to.equal(testSprite.x)
+        .to.equal(newValue);
+    });
+
+    it('aliases position.y to positionY', function() {
+      testSprite.position.y = 1;
+      expect(testSprite.position.y).to.equal(testSprite.y);
+      const newValue = 2;
+      testSprite.y = newValue;
+      expect(testSprite.position.y)
+        .to.equal(testSprite.y)
+        .to.equal(newValue);
+    });
+
+    it('aliases velocity.x to velocityX', function() {
+      testSprite.velocity.x = 1;
+      expect(testSprite.velocity.x).to.equal(testSprite.velocityX);
+      const newValue = 2;
+      testSprite.velocityX = newValue;
+      expect(testSprite.velocity.x)
+        .to.equal(testSprite.velocityX)
+        .to.equal(newValue);
+    });
+
+    it('aliases velocity.y to velocityY', function() {
+      testSprite.velocity.y = 1;
+      expect(testSprite.velocity.y).to.equal(testSprite.velocityY);
+      const newValue = 2;
+      testSprite.velocityY = newValue;
+      expect(testSprite.velocity.y)
+        .to.equal(testSprite.velocityY)
+        .to.equal(newValue);
+    });
+
+    it('aliases life to lifetime', function() {
+      testSprite.life = 1;
+      expect(testSprite.life).to.equal(testSprite.lifetime);
+      const newValue = 2;
+      testSprite.lifetime = newValue;
+      expect(testSprite.life)
+        .to.equal(testSprite.lifetime)
+        .to.equal(newValue);
+    });
+
+    it('aliases restitution to bounciness', function() {
+      testSprite.restitution = 1;
+      expect(testSprite.restitution).to.equal(testSprite.bounciness);
+      const newValue = 2;
+      testSprite.bounciness = newValue;
+      expect(testSprite.restitution)
+        .to.equal(testSprite.bounciness)
+        .to.equal(newValue);
+    });
+  });
+
+  describe('collisions with groups', function() {
+    let sprite, spriteTarget1, spriteTarget2, group;
+    let p5WrapperStateful, createStatefulGroup, createStatefulSprite;
+
+    beforeEach(function() {
+      p5WrapperStateful = createStatefulP5Wrapper();
+      createStatefulGroup = p5WrapperStateful.p5.createGroup.bind(
+        p5WrapperStateful.p5
+      );
+      createStatefulSprite = p5WrapperStateful.p5.createSprite.bind(
+        p5WrapperStateful.p5
+      );
+      spriteTarget1 = createStatefulSprite(0, 0, 100, 100);
+      spriteTarget2 = createStatefulSprite(400, 400, 100, 100);
+      sprite = createStatefulSprite(200, 200, 100, 100);
+      group = createStatefulGroup();
+      group.add(spriteTarget1);
+      group.add(spriteTarget2);
+    });
+
+    it('isTouching returns false when sprite touches nothing in the group', function() {
+      let result;
+      for (let i = 0; i < group.length + 1; i++) {
+        result = sprite.isTouching(group);
+      }
+      expect(result).to.equal(false);
+    });
+
+    it('isTouching returns true when sprite touches anything in the group', function() {
+      spriteTarget1.x = sprite.x;
+      spriteTarget1.y = sprite.y;
+
+      let result2;
+      for (let i = 0; i < group.length + 1; i++) {
+        result2 = sprite.isTouching(group);
+      }
+      expect(result2).to.equal(true);
+    });
+
+    it('collide returns false when sprite touches nothing in the group', function() {
+      let result;
+      for (let i = 0; i < group.length + 1; i++) {
+        result = sprite.collide(group);
+      }
+      expect(result).to.equal(false);
+    });
+
+    it('collide returns true when sprite touches anything in the group', function() {
+      spriteTarget1.x = sprite.x;
+      spriteTarget1.y = sprite.y;
+
+      let result2;
+      for (let i = 0; i < group.length + 1; i++) {
+        result2 = sprite.collide(group);
+      }
+      expect(result2).to.equal(true);
+    });
+
+    it('bounce returns false when sprite touches nothing in the group', function() {
+      let result;
+      for (let i = 0; i < group.length + 1; i++) {
+        result = sprite.bounce(group);
+      }
+      expect(result).to.equal(false);
+    });
+
+    it('bounce returns true when sprite touches anything in the group', function() {
+      spriteTarget1.x = sprite.x;
+      spriteTarget1.y = sprite.y;
+
+      let result2;
+      for (let i = 0; i < group.length + 1; i++) {
+        result2 = sprite.bounce(group);
+      }
+      expect(result2).to.equal(true);
+    });
+
+    it('bounceOff returns false when sprite touches nothing in the group', function() {
+      let result;
+      for (let i = 0; i < group.length + 1; i++) {
+        result = sprite.bounceOff(group);
+      }
+      expect(result).to.equal(false);
+    });
+
+    it('bounceOff returns true when sprite touches anything in the group', function() {
+      spriteTarget1.x = sprite.x;
+      spriteTarget1.y = sprite.y;
+
+      let result2;
+      for (let i = 0; i < group.length + 1; i++) {
+        result2 = sprite.bounceOff(group);
+      }
+      expect(result2).to.equal(true);
+    });
+
+    it('displace returns false when sprite touches nothing in the group', function() {
+      let result;
+      for (let i = 0; i < group.length + 1; i++) {
+        result = sprite.displace(group);
+      }
+      expect(result).to.equal(false);
+    });
+
+    it('displace returns true when sprite touches anything in the group', function() {
+      spriteTarget1.x = sprite.x;
+      spriteTarget1.y = sprite.y;
+
+      let result2;
+      for (let i = 0; i < group.length + 1; i++) {
+        result2 = sprite.displace(group);
+      }
+      expect(result2).to.equal(true);
+    });
+
+    it('overlap returns false when sprite touches nothing in the group', function() {
+      let result;
+      for (let i = 0; i < group.length + 1; i++) {
+        result = sprite.overlap(group);
+      }
+      expect(result).to.equal(false);
+    });
+
+    it('overlap returns true when sprite touches anything in the group', function() {
+      spriteTarget1.x = sprite.x;
+      spriteTarget1.y = sprite.y;
+
+      let result2;
+      for (let i = 0; i < group.length + 1; i++) {
+        result2 = sprite.overlap(group);
+      }
+      expect(result2).to.equal(true);
+    });
+  });
+
+  describe('sprite.collide(sprite)', function() {
+    var SIZE = 10;
+    var pInst;
+    var spriteA, spriteB;
+    var callCount, pairs;
+
+    function testCallback(a, b) {
+      callCount++;
+      pairs.push([a.name, b.name]);
+    }
+
+    function moveAToB(a, b) {
+      a.position.x = b.position.x;
+    }
+
+    beforeEach(function() {
+      pInst = new p5(function() {});
+      callCount = 0;
+      pairs = [];
+
+      function createTestSprite(letter, position) {
+        var sprite = pInst.createSprite(position, 0, SIZE, SIZE);
+        sprite.name = 'sprite' + letter;
+        return sprite;
+      }
+
+      spriteA = createTestSprite('A', 2 * SIZE);
+      spriteB = createTestSprite('B', 4 * SIZE);
+
+      // Assert initial test state:
+      // - Two total sprites
+      // - no two sprites overlap
+      expect(pInst.allSprites.length).to.equal(2);
+      pInst.allSprites.forEach(function(caller) {
+        pInst.allSprites.forEach(function(callee) {
+          expect(caller.overlap(callee)).to.be.false;
+        });
+      });
+    });
+
+    afterEach(function() {
+      pInst.remove();
+    });
+
+    it('false if sprites do not overlap', function() {
+      expect(spriteA.collide(spriteB)).to.be.false;
+      expect(spriteB.collide(spriteA)).to.be.false;
+    });
+
+    it('true if sprites overlap', function() {
+      moveAToB(spriteA, spriteB);
+      expect(spriteA.collide(spriteB)).to.be.true;
+
+      moveAToB(spriteA, spriteB);
+      expect(spriteB.collide(spriteA)).to.be.true;
+    });
+
+    it('calls callback once if sprites overlap', function() {
+      expect(callCount).to.equal(0);
+
+      moveAToB(spriteA, spriteB);
+      spriteA.collide(spriteB, testCallback);
+      expect(callCount).to.equal(1);
+
+      moveAToB(spriteA, spriteB);
+      spriteB.collide(spriteA, testCallback);
+      expect(callCount).to.equal(2);
+    });
+
+    it('does not call callback if sprites do not overlap', function() {
+      expect(callCount).to.equal(0);
+      spriteA.collide(spriteB, testCallback);
+      expect(callCount).to.equal(0);
+      spriteB.collide(spriteA, testCallback);
+      expect(callCount).to.equal(0);
+    });
+
+    describe('passes collider and collidee to callback', function() {
+      it('A-B', function() {
+        moveAToB(spriteA, spriteB);
+        spriteA.collide(spriteB, testCallback);
+        expect(pairs).to.deep.equal([[spriteA.name, spriteB.name]]);
+      });
+
+      it('B-A', function() {
+        moveAToB(spriteA, spriteB);
+        spriteB.collide(spriteA, testCallback);
+        expect(pairs).to.deep.equal([[spriteB.name, spriteA.name]]);
+      });
+    });
+
+    it('does not reposition either sprite when sprites do not overlap', function() {
+      var initialPositionA = spriteA.position.copy();
+      var initialPositionB = spriteB.position.copy();
+
+      spriteA.collide(spriteB);
+
+      expectVectorsAreClose(spriteA.position, initialPositionA);
+      expectVectorsAreClose(spriteB.position, initialPositionB);
+
+      spriteB.collide(spriteA);
+
+      expectVectorsAreClose(spriteA.position, initialPositionA);
+      expectVectorsAreClose(spriteB.position, initialPositionB);
+    });
+
+    describe('displaces the caller out of collision when sprites do overlap', function() {
+      it('to the left', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+
+        var expectedPositionA = spriteB.position.copy().add(-SIZE, 0);
+        var expectedPositionB = spriteB.position.copy();
+
+        spriteA.collide(spriteB);
+
+        expectVectorsAreClose(spriteA.position, expectedPositionA);
+        expectVectorsAreClose(spriteB.position, expectedPositionB);
+      });
+
+      it('to the right', function() {
+        spriteA.position.x = spriteB.position.x + 1;
+
+        var expectedPositionA = spriteB.position.copy().add(SIZE, 0);
+        var expectedPositionB = spriteB.position.copy();
+
+        spriteA.collide(spriteB);
+
+        expectVectorsAreClose(spriteA.position, expectedPositionA);
+        expectVectorsAreClose(spriteB.position, expectedPositionB);
+      });
+
+      it('caller and callee reversed', function() {
+        spriteA.position.x = spriteB.position.x + 1;
+
+        var expectedPositionA = spriteA.position.copy();
+        var expectedPositionB = spriteA.position.copy().add(-SIZE, 0);
+
+        spriteB.collide(spriteA);
+
+        expectVectorsAreClose(spriteA.position, expectedPositionA);
+        expectVectorsAreClose(spriteB.position, expectedPositionB);
+      });
+    });
+
+    it('does not change velocity of either sprite when sprites do not overlap', function() {
+      var initialVelocityA = spriteA.velocity.copy();
+      var initialVelocityB = spriteB.velocity.copy();
+
+      spriteA.collide(spriteB);
+
+      expectVectorsAreClose(spriteA.velocity, initialVelocityA);
+      expectVectorsAreClose(spriteB.velocity, initialVelocityB);
+
+      spriteB.collide(spriteA);
+
+      expectVectorsAreClose(spriteA.velocity, initialVelocityA);
+      expectVectorsAreClose(spriteB.velocity, initialVelocityB);
+    });
+
+    describe('matches caller velocity to callee velocity when sprites do overlap', function() {
+      it('when callee velocity is zero', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+        spriteA.velocity.x = 2;
+        spriteB.velocity.x = 0;
+
+        var expectedVelocityA = spriteB.velocity.copy();
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteA.collide(spriteB);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+
+      it('when callee velocity is nonzero', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+        spriteA.velocity.x = 2;
+        spriteB.velocity.x = -1;
+
+        var expectedVelocityA = spriteB.velocity.copy();
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteA.collide(spriteB);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+
+      it('only along x axis when only displaced along x axis', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+        spriteA.velocity.x = 2;
+        spriteA.velocity.y = 3;
+        spriteB.velocity.x = -1;
+
+        // Expect to change velocity only along X
+        var expectedVelocityA = spriteA.velocity.copy();
+        expectedVelocityA.x = spriteB.velocity.x;
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteA.collide(spriteB);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+
+      it('only along y axis when only displaced along y axis', function() {
+        spriteA.position.x = spriteB.position.x;
+        spriteA.position.y = spriteB.position.y + 1;
+        spriteA.velocity.x = 2;
+        spriteA.velocity.y = 3;
+        spriteB.velocity.x = -1;
+        spriteB.velocity.y = 1.5;
+
+        // Expect to change velocity only along Y
+        var expectedVelocityA = spriteA.velocity.copy();
+        expectedVelocityA.y = spriteB.velocity.y;
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteA.collide(spriteB);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+
+      it('caller and callee reversed', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+        spriteA.velocity.x = 2;
+        spriteB.velocity.x = -1;
+
+        var expectedVelocityA = spriteA.velocity.copy();
+        var expectedVelocityB = spriteA.velocity.copy();
+
+        spriteB.collide(spriteA);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+    });
+  });
+
+  describe('sprite.displace(sprite)', function() {
+    var SIZE = 10;
+    var pInst;
+    var spriteA, spriteB;
+    var callCount, pairs;
+
+    function testCallback(a, b) {
+      callCount++;
+      pairs.push([a.name, b.name]);
+    }
+
+    function moveAToB(a, b) {
+      a.position.x = b.position.x;
+    }
+
+    beforeEach(function() {
+      pInst = new p5(function() {});
+      callCount = 0;
+      pairs = [];
+
+      function createTestSprite(letter, position) {
+        var sprite = pInst.createSprite(position, 0, SIZE, SIZE);
+        sprite.name = 'sprite' + letter;
+        return sprite;
+      }
+
+      spriteA = createTestSprite('A', 2 * SIZE);
+      spriteB = createTestSprite('B', 4 * SIZE);
+
+      // Assert initial test state:
+      // - Two total sprites
+      // - no two sprites overlap
+      expect(pInst.allSprites.length).to.equal(2);
+      pInst.allSprites.forEach(function(caller) {
+        pInst.allSprites.forEach(function(callee) {
+          expect(caller.overlap(callee)).to.be.false;
+        });
+      });
+    });
+
+    afterEach(function() {
+      pInst.remove();
+    });
+
+    it('false if sprites do not overlap', function() {
+      expect(spriteA.displace(spriteB)).to.be.false;
+      expect(spriteB.displace(spriteA)).to.be.false;
+    });
+
+    it('true if sprites overlap', function() {
+      moveAToB(spriteA, spriteB);
+      expect(spriteA.displace(spriteB)).to.be.true;
+
+      moveAToB(spriteA, spriteB);
+      expect(spriteB.displace(spriteA)).to.be.true;
+    });
+
+    it('calls callback once if sprites overlap', function() {
+      expect(callCount).to.equal(0);
+
+      moveAToB(spriteA, spriteB);
+      spriteA.displace(spriteB, testCallback);
+      expect(callCount).to.equal(1);
+
+      moveAToB(spriteA, spriteB);
+      spriteB.displace(spriteA, testCallback);
+      expect(callCount).to.equal(2);
+    });
+
+    it('does not call callback if sprites do not overlap', function() {
+      expect(callCount).to.equal(0);
+      spriteA.displace(spriteB, testCallback);
+      expect(callCount).to.equal(0);
+      spriteB.displace(spriteA, testCallback);
+      expect(callCount).to.equal(0);
+    });
+
+    describe('passes collider and collidee to callback', function() {
+      it('A-B', function() {
+        moveAToB(spriteA, spriteB);
+        spriteA.displace(spriteB, testCallback);
+        expect(pairs).to.deep.equal([[spriteA.name, spriteB.name]]);
+      });
+
+      it('B-A', function() {
+        moveAToB(spriteA, spriteB);
+        spriteB.displace(spriteA, testCallback);
+        expect(pairs).to.deep.equal([[spriteB.name, spriteA.name]]);
+      });
+    });
+
+    it('does not reposition either sprite when sprites do not overlap', function() {
+      var initialPositionA = spriteA.position.copy();
+      var initialPositionB = spriteB.position.copy();
+
+      spriteA.displace(spriteB);
+
+      expectVectorsAreClose(spriteA.position, initialPositionA);
+      expectVectorsAreClose(spriteB.position, initialPositionB);
+
+      spriteB.displace(spriteA);
+
+      expectVectorsAreClose(spriteA.position, initialPositionA);
+      expectVectorsAreClose(spriteB.position, initialPositionB);
+    });
+
+    describe('displaces the callee out of collision when sprites do overlap', function() {
+      it('to the left', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+
+        var expectedPositionA = spriteA.position.copy();
+        var expectedPositionB = spriteA.position.copy().add(SIZE, 0);
+
+        spriteA.displace(spriteB);
+
+        expectVectorsAreClose(spriteA.position, expectedPositionA);
+        expectVectorsAreClose(spriteB.position, expectedPositionB);
+      });
+
+      it('to the right', function() {
+        spriteA.position.x = spriteB.position.x + 1;
+
+        var expectedPositionA = spriteA.position.copy();
+        var expectedPositionB = spriteA.position.copy().add(-SIZE, 0);
+
+        spriteA.displace(spriteB);
+
+        expectVectorsAreClose(spriteA.position, expectedPositionA);
+        expectVectorsAreClose(spriteB.position, expectedPositionB);
+      });
+
+      it('caller and callee reversed', function() {
+        spriteA.position.x = spriteB.position.x + 1;
+
+        var expectedPositionA = spriteB.position.copy().add(SIZE, 0);
+        var expectedPositionB = spriteB.position.copy();
+
+        spriteB.displace(spriteA);
+
+        expectVectorsAreClose(spriteA.position, expectedPositionA);
+        expectVectorsAreClose(spriteB.position, expectedPositionB);
+      });
+    });
+
+    it('does not change velocity of either sprite when sprites do not overlap', function() {
+      var initialVelocityA = spriteA.velocity.copy();
+      var initialVelocityB = spriteB.velocity.copy();
+
+      spriteA.displace(spriteB);
+
+      expectVectorsAreClose(spriteA.velocity, initialVelocityA);
+      expectVectorsAreClose(spriteB.velocity, initialVelocityB);
+
+      spriteB.displace(spriteA);
+
+      expectVectorsAreClose(spriteA.velocity, initialVelocityA);
+      expectVectorsAreClose(spriteB.velocity, initialVelocityB);
+    });
+
+    describe('does not change callee velocity', function() {
+      it('when caller velocity is zero', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+        spriteA.velocity.x = 0;
+        spriteB.velocity.x = 2;
+
+        var expectedVelocityA = spriteA.velocity.copy();
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteA.displace(spriteB);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+
+      it('when caller velocity is nonzero', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+        spriteA.velocity.x = 2;
+        spriteB.velocity.x = -1;
+
+        var expectedVelocityA = spriteA.velocity.copy();
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteA.displace(spriteB);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+
+      it('when only displaced along x axis', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+        spriteA.velocity.x = 2;
+        spriteA.velocity.y = 0.5;
+        spriteB.velocity.x = -1;
+        spriteB.velocity.y = 3;
+
+        var expectedVelocityA = spriteA.velocity.copy();
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteA.displace(spriteB);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+
+      it('when only displaced along y axis', function() {
+        spriteA.position.x = spriteB.position.x;
+        spriteA.position.y = spriteB.position.y + 1;
+        spriteA.velocity.x = 2;
+        spriteA.velocity.y = 3;
+        spriteB.velocity.x = -1;
+        spriteB.velocity.y = 1.5;
+
+        var expectedVelocityA = spriteA.velocity.copy();
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteA.displace(spriteB);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+
+      it('caller and callee reversed', function() {
+        spriteA.position.x = spriteB.position.x - 1;
+        spriteA.velocity.x = 2;
+        spriteB.velocity.x = -1;
+
+        var expectedVelocityA = spriteA.velocity.copy();
+        var expectedVelocityB = spriteB.velocity.copy();
+
+        spriteB.displace(spriteA);
+
+        expectVectorsAreClose(spriteA.velocity, expectedVelocityA);
+        expectVectorsAreClose(spriteB.velocity, expectedVelocityB);
+      });
+    });
+  });
+
+  function expectVectorsAreClose(vA, vB) {
+    var failMsg =
+      'Expected <' +
+      vA.x +
+      ', ' +
+      vA.y +
+      '> to equal <' +
+      vB.x +
+      ', ' +
+      vB.y +
+      '>';
+    expect(vA.x).to.be.closeTo(vB.x, 0.00001, failMsg);
+    expect(vA.y).to.be.closeTo(vB.y, 0.00001, failMsg);
+  }
+});
